feat: determine CI provider via env var heuristics
We are trying to collect data about using CI or interactive
terminal so we know what impacts future local builds will have
on developer workflows.

Using environment variable heuristics from watson/ci-info
we add "ci" or "terminal" to the user-agent string for
outbound API requests.

We can detect the CI provider, so in the future we could
also track that information; currently, we do not.

Signed-off-by: Chris Goller <goller@gmail.com>