feat(load): use docker attach to proxy docker pull
When the docker daemon was remote, it could not always
reach the depot CLI registry.

In some environments like CircleCI the communication to a running
container was limited to just the docker API.

Docker attach provides a TCP connection that proxies a container's
STDIN and STDOUT.

This uses a serialized, framed packet format to receive blob requests
from the remote container's STDOUT and send blob responses to the
container's STDIN.

The protocol is binary and requires TTY to not be "true."

The framing protocol consists of a session id, data length, and data.

Regarding performance, it seems similar as before; sometimes slower
and sometimes faster.

Signed-off-by: Chris Goller <goller@gmail.com>