perf: remove extra solve used for capability discovery
To determine if the buildkit server supported S3 and gha
caching, an entire build cycle was used simply to run
"ping."

The build ran five gRPC requests of `ListWorkers`, `Status`, `Session`,
`Solve`, and `Ping`.

Ultimately, it was the `Ping` that matters as it returns the capabilities
of the server.

Of the dozens of capabilities the client looks two for caching (s3 and gha)
and attestations.

Depot supports those three and thus we can remove this entire cycle.

Incidentally, this first build happens to start a garbage collection
that runs during the real second build.

Signed-off-by: Chris Goller <goller@gmail.com>