feat: add manifest to metadata-file output
In order to alert CI pipelines to a large increase in layer size
the image manifest is attached to the metadata-file.

This assumes that an image has been created.

There are some bugs in buildx bake's manifest with multi-platforms;
specifically, the format does not really support multi-platform, but
has suffixes for JSON keys for a few blessed fields.

Signed-off-by: Chris Goller <goller@gmail.com>