feat(load): wait for proxy ready signal before pull
Heavily loaded machines may be slow to start the HTTP registry server.
Now, we wait to receive the version over STDERR before instructing
docker to pull the image.

The current wait-time is up to 10 seconds; if we do not receive
the version number before that time, we try to pull anyway
with failures causing our standard use of docker tar load.

Signed-off-by: Chris Goller <goller@gmail.com>