refactor(load): use manifests and configs from solve response
Previously, manifests and configs were sent as annotations
on the image index. This caused failures when pushing to GCR
and oci-mediatypes was false.

Now a new export image option, `depot.export.image.version` has
been added causing the builder service to return the manifests and
configs in the solve response at the key `depot.export.image.version`.

This is backwards compatible for older builder services.

Signed-off-by: Chris Goller <goller@gmail.com>