feat(buildx): bootstrap driver by creating container
docker compose has a race condition with docker buildx
that causes more than one container with the same image
name to be created.

This attempts to download the newest CLI container during
the setup of the driver.

Once the upstream issue (buildx #2000) is released we
can remove these extra steps.

Signed-off-by: Chris Goller <goller@gmail.com>