fix: use unmodifed `image.name` in metadata file
In multi-platform builds buildx strips the image tag
and uses `push-by-digest` for the manifest. This is to
prevent races between the builds to set the user specified
tag.

However, this caused the downstream image.name to be the
tag-less name.

This fix adds back the user's specified image names to fix
the `image.name` field.

Signed-off-by: Chris Goller <goller@gmail.com>